job_aggregator/requirements.txt
requests
sqlalchemy
psycopg2-binary
sentence-transformers
numpy
scikit-learn
python-dotenv
